func: split into src/functions.py
